penumbra_sdk_shielded_pool
Module fmd

Modules

state_key

Structs

MetaParameters

Meta parameters governing how FMD parameters change.

Parameters

SlidingWindow

A struct holding params for the sliding window algorithm

Enums

MetaParametersAlgorithm

MetaParametersAlgorithmState

If any, the current state for the algorithm we’re using.

Constants

FMD_GRACE_PERIOD_BLOCKS_DEFAULT

How long users have to switch to updated parameters.

Functions

should_update_fmd_params